requirements.txt
cryptography>=40.0.0
Pillow>=9.0.0
piexif>=1.1.3
pikepdf>=6.0.0    
python-magic>=0.4.27